// src/services/sited_io/commerce/v1/offer_pb.ts
export enum OfferType {
  /**
   * @generated from enum value: OFFER_TYPE_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: OFFER_TYPE_PHYSICAL = 1;
   */
  PHYSICAL = 1,

  /**
   * @generated from enum value: OFFER_TYPE_DIGITAL = 2;
   */
  DIGITAL = 2,

  /**
   * @generated from enum value: OFFER_TYPE_SERVICE = 3;
   */
  SERVICE = 3,
}
// ...
export class OfferResponse extends Message<OfferResponse> {
  /**
   * @generated from field: string offer_id = 1;
   */
  offerId = "";

  /**
   * @generated from field: string shop_id = 2;
   */
  shopId = "";

  /**
   * @generated from field: string shop_name = 3;
   */
  shopName = "";

  /**
   * @generated from field: string shop_slug = 4;
   */
  shopSlug = "";

  /**
   * @generated from field: optional string shop_domain = 5;
   */
  shopDomain?: string;

  /**
   * @generated from field: string user_id = 6;
   */
  userId = "";

  /**
   * @generated from field: int64 created_at = 7;
   */
  createdAt = protoInt64.zero;

  /**
   * @generated from field: int64 updated_at = 8;
   */
  updatedAt = protoInt64.zero;

  /**
   * @generated from field: string name = 9;
   */
  name = "";

  /**
   * @generated from field: string description = 10;
   */
  description = "";

  /**
   * @generated from field: bool is_active = 11;
   */
  isActive = false;

  /**
   * @generated from field: bool is_featured = 12;
   */
  isFeatured = false;

  /**
   * @generated from field: sited_io.commerce.v1.OfferType type = 13;
   */
  type = OfferType.UNSPECIFIED;

  /**
   * @generated from field: repeated sited_io.commerce.v1.OfferImageResponse images = 14;
   */
  images: OfferImageResponse[] = [];

  /**
   * @generated from field: optional sited_io.commerce.v1.Price price = 15;
   */
  price?: Price;

  constructor(data?: PartialMessage<OfferResponse>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "sited_io.commerce.v1.OfferResponse";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "offer_id", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 2, name: "shop_id", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 3, name: "shop_name", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 4, name: "shop_slug", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 5, name: "shop_domain", kind: "scalar", T: 9 /* ScalarType.STRING */, opt: true },
    { no: 6, name: "user_id", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 7, name: "created_at", kind: "scalar", T: 3 /* ScalarType.INT64 */ },
    { no: 8, name: "updated_at", kind: "scalar", T: 3 /* ScalarType.INT64 */ },
    { no: 9, name: "name", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 10, name: "description", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 11, name: "is_active", kind: "scalar", T: 8 /* ScalarType.BOOL */ },
    { no: 12, name: "is_featured", kind: "scalar", T: 8 /* ScalarType.BOOL */ },
    { no: 13, name: "type", kind: "enum", T: proto3.getEnumType(OfferType) },
    { no: 14, name: "images", kind: "message", T: OfferImageResponse, repeated: true },
    { no: 15, name: "price", kind: "message", T: Price, opt: true },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): OfferResponse {
    return new OfferResponse().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): OfferResponse {
    return new OfferResponse().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): OfferResponse {
    return new OfferResponse().fromJsonString(jsonString, options);
  }

  static equals(a: OfferResponse | PlainMessage<OfferResponse> | undefined, b: OfferResponse | PlainMessage<OfferResponse> | undefined): boolean {
    return proto3.util.equals(OfferResponse, a, b);
  }
}

/**
 * @generated from message sited_io.commerce.v1.OfferImageResponse
 */
export class OfferImageResponse extends Message<OfferImageResponse> {
  /**
   * @generated from field: string offer_image_id = 1;
   */
  offerImageId = "";

  /**
   * @generated from field: string image_url = 2;
   */
  imageUrl = "";

  /**
   * @generated from field: int64 ordering = 3;
   */
  ordering = protoInt64.zero;

  constructor(data?: PartialMessage<OfferImageResponse>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "sited_io.commerce.v1.OfferImageResponse";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "offer_image_id", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 2, name: "image_url", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 3, name: "ordering", kind: "scalar", T: 3 /* ScalarType.INT64 */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): OfferImageResponse {
    return new OfferImageResponse().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): OfferImageResponse {
    return new OfferImageResponse().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): OfferImageResponse {
    return new OfferImageResponse().fromJsonString(jsonString, options);
  }

  static equals(a: OfferImageResponse | PlainMessage<OfferImageResponse> | undefined, b: OfferImageResponse | PlainMessage<OfferImageResponse> | undefined): boolean {
    return proto3.util.equals(OfferImageResponse, a, b);
  }
}
// ...
export class PutPriceToOfferRequest extends Message<PutPriceToOfferRequest> {
  /**
   * @generated from field: string offer_id = 1;
   */
  offerId = "";

  /**
   * @generated from field: sited_io.commerce.v1.Price price = 2;
   */
  price?: Price;

  constructor(data?: PartialMessage<PutPriceToOfferRequest>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "sited_io.commerce.v1.PutPriceToOfferRequest";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "offer_id", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 2, name: "price", kind: "message", T: Price },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): PutPriceToOfferRequest {
    return new PutPriceToOfferRequest().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): PutPriceToOfferRequest {
    return new PutPriceToOfferRequest().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): PutPriceToOfferRequest {
    return new PutPriceToOfferRequest().fromJsonString(jsonString, options);
  }

  static equals(a: PutPriceToOfferRequest | PlainMessage<PutPriceToOfferRequest> | undefined, b: PutPriceToOfferRequest | PlainMessage<PutPriceToOfferRequest> | undefined): boolean {
    return proto3.util.equals(PutPriceToOfferRequest, a, b);
  }
}